TAS-34 fix scenario edit form work
